DayTwelve
DayTwelve includes the use of arrow functions. It recives one parameter, an array called nums. The function must iterate through the array performing one of the following actions on each element:
If the element is even, multiply the element by 2
If the element is odd, multiply the element by 3
And return the modified array