DayTwenty
DayTwenty includes decoding a matrix script that is a N x M grid of strings, and consists of alphanumeric characters, spaces and symbols (!,@,#,$,%,&).
It reads each column, selects only the alphanumeric characters and connects them.
If there are symbols or spaces between two alphanumeric characters of the decoded script, then they get replaced with a single space ' ' for better readability.