DayTwentyOne
DayTwentyOne includes debugging the existing code to get the correct output.
The correct output was achieved by adding a virtual destructor to the base class, HotelRoom, for proper clean up.
The code provided was dynamically creating instances of the base class, but in C++ when allocating memory dynamically using 'new' that memory has to later be freed to avoid memory leaks and undefined behavior.